More work on component_union()
